requirements.txt
# Core ML
scikit-learn==1.3.2
xgboost==2.0.2
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Data fetching
requests==2.31.0
aiohttp==3.10.11

# Real-time
asyncio-mqtt==0.16.1

# UI
streamlit==1.29.0
plotly==5.18.0

# Utils
pyyaml==6.0.1
python-dotenv==1.0.0
joblib==1.3.2

# Database
aiosqlite==0.19.0

# Optional (Telegram)
python-telegram-bot==20.7
